fix: all example page migration
