+++ app/examples/kanban/page.tsx
@@ -1,0 +1,5 @@
+import KanbanPage from './kanban-page';
+
+export default async function Page() {
+  return <KanbanPage />;
+}
