+++ app/examples/multi-cursor/page.tsx
@@ -1,0 +1,5 @@
+import MultiCursorPage from './multi-cursor-page';
+
+export default async function Page() {
+  return <MultiCursorPage />;
+}
