+++ app/examples/profile-stack/page.tsx
@@ -1,0 +1,5 @@
+import ProfileStackPage from './profile-stack-page';
+
+export default async function Page() {
+  return <ProfileStackPage />;
+}
