+++ app/examples/quill/page.tsx
@@ -1,0 +1,5 @@
+import QuillPage from './quill-page';
+
+export default async function Page() {
+  return <QuillPage />;
+}
