+++ app/examples/quill/quill-page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { NextPage } from 'next';
 import Head from 'next/head';
 import { ExampleLayout } from '@/components';
@@ -9,11 +11,12 @@ import {
   EXAMPLE_CODE_URL,
   EXAMPLE_PREVIEW_URL,
 } from '@/components/exampleView';
-import { FILE_INFO } from '@/examples/profile-stack/fileInfo';
+import { FILE_INFO } from '@/examples/vanilla-quill/fileInfo';
 
-const exampleKey = 'profile-stack';
-const exampleTitle = 'Profile Stack';
-const ProfileStackExampleView: NextPage = () => {
+const exampleKey = 'vanilla-quill';
+const exampleTitle = 'Quill';
+
+const QuillExampleView: NextPage = () => {
   return (
     <ExampleLayout breadcrumbTitle={exampleTitle}>
       {() => (
@@ -31,13 +34,24 @@ const ProfileStackExampleView: NextPage = () => {
               <Sidebar.TabsPanel value="code">
                 <Sidebar.GuideTitle>{exampleTitle}</Sidebar.GuideTitle>
                 <Sidebar.GuideDescription>
-                  The profile stack shows the list of users currently accessing the Document. Try adding and deleting
-                  users to see how the profile stack changes.
+                  This demo shows the real-time collaborative version of the{' '}
+                  <a href="https://quilljs.com/" className="link" target="_blank" rel="noreferrer">
+                    Quill
+                  </a>{' '}
+                  editor with{' '}
+                  <a href="https://yorkie.dev/" className="link" target="_blank" rel="noreferrer">
+                    Yorkie
+                  </a>{' '}
+                  and{' '}
+                  <a href="https://vitejs.dev/" className="link" target="_blank" rel="noreferrer">
+                    Vite
+                  </a>
+                  .
                 </Sidebar.GuideDescription>
                 <ProjectCodes
                   files={FILE_INFO}
-                  activeFile="/main.js"
-                  ignoreFiles={[...COMMON_IGNORE_FILES, '.env', 'vite.config.js']}
+                  activeFile="/src/main.ts"
+                  ignoreFiles={[...COMMON_IGNORE_FILES, '.env', 'vite.config.js', '/src/vite-env.d.ts']}
                 />
               </Sidebar.TabsPanel>
               <Sidebar.Bottom codeURL={EXAMPLE_CODE_URL + exampleKey} />
@@ -48,12 +62,10 @@ const ProfileStackExampleView: NextPage = () => {
             apiKey={process.env.NEXT_PUBLIC_EXAMPLES_API_KEY || ''}
             documentKey={exampleKey}
             iframeURL={EXAMPLE_PREVIEW_URL + exampleKey}
-            userMaxCount={6}
           />
         </>
       )}
     </ExampleLayout>
   );
 };
-
-export default ProfileStackExampleView;
+export default QuillExampleView;
