+++ app/examples/simultaneous-cursors/page.tsx
@@ -1,0 +1,5 @@
+import SimultaneousCursorsPage from './simultaneous-cursors-page';
+
+export default async function Page() {
+  return <SimultaneousCursorsPage />;
+}
