+++ app/examples/tldraw/page.tsx
@@ -1,0 +1,5 @@
+import TldrawPage from './tldraw-page';
+
+export default async function Page() {
+  return <TldrawPage />;
+}
