+++ app/examples/todomvc/page.tsx
@@ -1,0 +1,5 @@
+import TodomvcPage from './todomvc-page';
+
+export default async function Page() {
+  return <TodomvcPage />;
+}
