+++ app/examples/todomvc/todomvc-page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { ExampleLayout } from '@/components';
 import {
   BasicExampleView,
