+++ app/examples/webtoons/page.tsx
@@ -1,0 +1,5 @@
+import WebtoonsPage from './webtoons-page';
+
+export default async function Page() {
+  return <WebtoonsPage />;
+}
